Rename clusterGraph.ts -> cytoscapeData.ts
